# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/salary_slip_lines:
    get:
      tags:
        - SalarySlipLines
      summary: List SalarySlipLines
      description: Get a paginated list of SalarySlipLines with optional filtering and sorting
      operationId: listSalarySlipLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalarySlipLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalarySlipLines
      summary: Create SalarySlipLine
      description: Create a new SalarySlipLine
      operationId: createSalarySlipLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalarySlipLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalarySlipLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/salary_structures:
    get:
      tags:
        - SalaryStructures
      summary: List SalaryStructures
      description: Get a paginated list of SalaryStructures with optional filtering and sorting
      operationId: listSalaryStructures
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryStructureList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalaryStructures
      summary: Create SalaryStructure
      description: Create a new SalaryStructure
      operationId: createSalaryStructure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalaryStructureRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryStructure'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    SalarySlipLine:
      type: object
      required:
        - id
        - salary_slip_id
        - company_id
        - name
        - component_type
        - is_statutory
        - amount
        - gl_account_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        salary_slip_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        component_type:
          type: string
        is_statutory:
          type: boolean
        amount:
          type: string
        gl_account_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateSalarySlipLineRequest:
      type: object
      required:
        - salary_slip_id
        - company_id
        - name
        - component_type
        - is_statutory
        - amount
        - gl_account_id
        - metadata
      properties:
        salary_slip_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        component_type:
          type: string
        is_statutory:
          type: boolean
        amount:
          type: string
        gl_account_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    SalarySlipLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalarySlipLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    SalaryStructure:
      type: object
      required:
        - id
        - company_id
        - name
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        status:
          type: string
        metadata:
          type: object
...
    CreateSalaryStructureRequest:
      type: object
      required:
        - company_id
        - name
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        status:
          type: string
        metadata:
          type: object
...
    SalaryStructureList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalaryStructure'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []